pricing model refactor
