+++ elfpy/markets/hyperdrive/hyperdrive_market.py
@@ -8,9 +8,9 @@ from dataclasses import dataclass, field
 import elfpy
 import elfpy.agents.wallet as wallet
 import elfpy.errors.errors as errors
-import elfpy.markets.base as base_market
+import elfpy.markets.base_market as base_market
 import elfpy.markets.hyperdrive.hyperdrive_actions as hyperdrive_actions
-import elfpy.pricing_models.hyperdrive as hyperdrive_pm
+import elfpy.markets.hyperdrive.hyperdrive_pricing_model as hyperdrive_pm
 import elfpy.time as time
 import elfpy.types as types
 import elfpy.utils.price as price_utils
