+++ elfpy/markets/hyperdrive/hyperdrive_pricing_model.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 
 import elfpy.agents.agent_trade_result as agent_trade_result
 import elfpy.markets.hyperdrive.market_action_result as market_action_result
-import elfpy.pricing_models.trades as trades
-import elfpy.pricing_models.yieldspace as yieldspace_pm
+import elfpy.markets.hyperdrive.yieldspace_pricing_model as yieldspace_pm
+import elfpy.markets.trades as trades
 import elfpy.time as time
 import elfpy.types as types
 from elfpy.math import FixedPoint
