--- elfpy/pricing_models/__init__.py
@@ -1,1 +1,0 @@
-"""Stateless pricing models for computing trade outcomes"""
