+++ elfpy/utils/sim_utils.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from elfpy.agents.policies.init_lp import InitializeLiquidityAgent
 
 import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
-import elfpy.pricing_models.hyperdrive as hyperdrive_pm
+import elfpy.markets.hyperdrive.hyperdrive_pricing_model as hyperdrive_pm
 import elfpy.simulators as simulators
 import elfpy.time as time
 from elfpy.math import FixedPoint
