+++ tests/solidity/test_close_long.py
@@ -4,7 +4,7 @@ import unittest
 import elfpy.agents.agent as elf_agent
 from elfpy.markets.hyperdrive.checkpoint import Checkpoint
 import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
-import elfpy.pricing_models.hyperdrive as hyperdrive_pm
+import elfpy.markets.hyperdrive.hyperdrive_pricing_model as hyperdrive_pm
 import elfpy.time as time
 from elfpy.time.time import TimeUnit
 import elfpy.types as types
