+++ tests/solidity/test_close_short.py
@@ -4,8 +4,8 @@ import unittest
 import elfpy.agents.agent as elf_agent
 from elfpy.markets.hyperdrive.checkpoint import Checkpoint
 import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
-import elfpy.pricing_models.hyperdrive as hyperdrive_pm
-import elfpy.pricing_models.yieldspace as yieldspace_pm
+import elfpy.markets.hyperdrive.hyperdrive_pricing_model as hyperdrive_pm
+import elfpy.markets.hyperdrive.yieldspace_pricing_model as yieldspace_pm
 import elfpy.time as time
 import elfpy.types as types
 from elfpy.time.time import StretchedTime
